+++ lib/wanikani/configuration.rb
@@ -1,14 +1,13 @@
 module Wanikani
   class Configuration
-    attr_accessor :api_version, :api_revision, :api_key
+    attr_accessor :api_revision, :api_key
 
     def initialize(configuration)
-      @api_version = configuration.fetch(:api_version)
+      @api_revision = configuration.fetch(:api_revision)
     end
 
     def to_hash
       {
-        api_version: api_version,
         api_revision: api_revision,
         api_key: api_key
       }