+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # WaniKani gem - Changelog
 
+### 3.0.0
+- **Breaking**: The wanikani gem was completely revamped to support Wanikani's V2 API.
+. Please read the [documentation](https://github.com/dennmart/wanikani-gem) for more information.
+
+
 ### 2.0.0
 - **Breaking**: The wanikani gem was completely revamped to use a per-instance client. Please read the [documentation](https://github.com/dennmart/wanikani-gem) for more information.